Minor Cygwin fixes
git-svn-id: http://svn.csrri.iit.edu/mx/trunk@516 42547f41-83fc-0310-9aef-9668e019ab9e

+++ libMx/mx_record.c
@@ -1287,7 +1287,7 @@ mx_read_database_private( MX_RECORD *record_list_head,
 
 		MX_DEBUG( 2,
 		("%s invoked, is_array = %d, num_lines = %ld, array_ptr = %p",
-			fname, db_source->is_array,
+			fname, (int) db_source->is_array,
 			db_source->num_lines, db_source->array_ptr ));
 
 		mxp_readline = mxp_readline_from_array;
@@ -1296,7 +1296,7 @@ mx_read_database_private( MX_RECORD *record_list_head,
 
 		MX_DEBUG( 2,
 		("%s invoked, is_array = %d, filename = '%s'",
-			fname, db_source->is_array, db_source->filename ));
+		    fname, (int) db_source->is_array, db_source->filename ));
 
 		mxp_readline = mxp_readline_from_file;
